requirements.txt
gunicorn
google-generativeai==0.8.3
python-dotenv==1.0.1
chromadb==1.0.9
langchain==0.3.11
langchain-community==0.3.11
langchain-google-genai==2.0.10
langchain-chroma==0.2.2
flask==3.0.3
markdown==3.6 